pyproject.toml
[tool.poetry]
name = "gym-microrts"
version = "0.1.3"
description = ""
authors = ["Costa Huang <costa.huang@outlook.com>"]
packages = [
    { include = "gym_microrts" },
]
include = ["microrts/**/*", "README.md"]


[tool.poetry.dependencies]
python = "^3.7"
gym = "^0.17.3"
JPype1 = "^1.0.2"
dacite = "^1.5.1"
Pillow = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
